Use uint8_t for numGears
Rest of the bytes seem kinda uninitialized.

+++ RTHandlingEditor/HandlingItem.h
@@ -19,7 +19,7 @@ struct CHandlingDataItem {
     float vecInertiaMultiplierY;
     float vecInertiaMultiplierZ;
     float fDriveBiasFront;
-    int   nInitialDriveGears;
+    uint8_t nInitialDriveGears;
     float fInitialDriveForce;
     float fDriveInertia;
     float fClutchChangeRateScaleUpShift;
